requirements.txt
python-dotenv~=1.0.1
mcstatus==11.1.1
aiogram~=3.17.0